Added interface base class. Migrated stoll-werth code to interface. Refactor models ready for interface based models.
