+++ surrogates/interfaces/__init__.py
@@ -1,0 +1,3 @@
+from .interfaces import EvaluationInterface
+
+__all__ = [EvaluationInterface]
