+++ surrogates/tests/models/test_trained.py
@@ -17,58 +17,42 @@ def test_gaussian_process_no_noise():
 
     model.add_training_data(
         numpy.array([[0.0, 0.0]]),
-        {
-            "liquid_density": numpy.array([1.0]),
-            "vapor_pressure": numpy.array([2.0]),
-            "surface_tension": numpy.array([3.0]),
-        },
-        {
-            "liquid_density": numpy.array([0.0]),
-            "vapor_pressure": numpy.array([0.0]),
-            "surface_tension": numpy.array([0.0]),
-        },
+        {"a": numpy.array([1.0]), "b": numpy.array([2.0]), "c": numpy.array([3.0])},
+        {"a": numpy.array([0.0]), "b": numpy.array([0.0]), "c": numpy.array([0.0])},
     )
 
-    values, uncertainties = model.evaluate(numpy.array([[0.0, 0.0]]))
+    values, uncertainties = model.evaluate(["a", "b", "c"], numpy.array([[0.0, 0.0]]))
 
-    assert numpy.isclose(values["liquid_density"], 1.0)
-    assert numpy.isclose(values["vapor_pressure"], 2.0)
-    assert numpy.isclose(values["surface_tension"], 3.0)
+    assert numpy.isclose(values["a"], 1.0)
+    assert numpy.isclose(values["b"], 2.0)
+    assert numpy.isclose(values["c"], 3.0)
 
-    assert numpy.isclose(uncertainties["liquid_density"], 0.0)
-    assert numpy.isclose(uncertainties["vapor_pressure"], 0.0)
-    assert numpy.isclose(uncertainties["surface_tension"], 0.0)
+    assert numpy.isclose(uncertainties["a"], 0.0)
+    assert numpy.isclose(uncertainties["b"], 0.0)
+    assert numpy.isclose(uncertainties["c"], 0.0)
 
     model.add_training_data(
         numpy.array([[1.0, 1.0]]),
-        {
-            "liquid_density": numpy.array([4.0]),
-            "vapor_pressure": numpy.array([5.0]),
-            "surface_tension": numpy.array([6.0]),
-        },
-        {
-            "liquid_density": numpy.array([0.0]),
-            "vapor_pressure": numpy.array([0.0]),
-            "surface_tension": numpy.array([0.0]),
-        },
+        {"a": numpy.array([4.0]), "b": numpy.array([5.0]), "c": numpy.array([6.0])},
+        {"a": numpy.array([0.0]), "b": numpy.array([0.0]), "c": numpy.array([0.0])},
     )
 
-    values, uncertainties = model.evaluate(numpy.array([[0.0, 0.0]]))
+    values, uncertainties = model.evaluate(["a", "b", "c"], numpy.array([[0.0, 0.0]]))
 
-    assert numpy.isclose(values["liquid_density"], 1.0)
-    assert numpy.isclose(values["vapor_pressure"], 2.0)
-    assert numpy.isclose(values["surface_tension"], 3.0)
+    assert numpy.isclose(values["a"], 1.0)
+    assert numpy.isclose(values["b"], 2.0)
+    assert numpy.isclose(values["c"], 3.0)
 
-    assert numpy.isclose(uncertainties["liquid_density"], 0.0)
-    assert numpy.isclose(uncertainties["vapor_pressure"], 0.0)
-    assert numpy.isclose(uncertainties["surface_tension"], 0.0)
+    assert numpy.isclose(uncertainties["a"], 0.0)
+    assert numpy.isclose(uncertainties["b"], 0.0)
+    assert numpy.isclose(uncertainties["c"], 0.0)
 
-    values, uncertainties = model.evaluate(numpy.array([[1.0, 1.0]]))
+    values, uncertainties = model.evaluate(["a", "b", "c"], numpy.array([[1.0, 1.0]]))
 
-    assert numpy.isclose(values["liquid_density"], 4.0)
-    assert numpy.isclose(values["vapor_pressure"], 5.0)
-    assert numpy.isclose(values["surface_tension"], 6.0)
+    assert numpy.isclose(values["a"], 4.0)
+    assert numpy.isclose(values["b"], 5.0)
+    assert numpy.isclose(values["c"], 6.0)
 
-    assert numpy.isclose(uncertainties["liquid_density"], 0.0)
-    assert numpy.isclose(uncertainties["vapor_pressure"], 0.0)
-    assert numpy.isclose(uncertainties["surface_tension"], 0.0)
+    assert numpy.isclose(uncertainties["a"], 0.0)
+    assert numpy.isclose(uncertainties["b"], 0.0)
+    assert numpy.isclose(uncertainties["c"], 0.0)
